gcf/requirements.txt
google-cloud-texttospeech==2.2.0
boto3==1.17.22
